merquemos/requirements.txt
adjacent==1.0.0
astroid==1.6.5
backports.functools-lru-cache==1.5
cairocffi==0.9.0
CairoSVG==1.0.22
cent==2.1.0
certifi==2017.4.17
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
contextlib2==0.5.5
coreapi==2.3.1
coreschema==0.0.4
cssselect==1.0.3
cssselect2==0.2.1
diff-match-patch==20121119
Django==1.11.2
django-allauth==0.32.0
django-easy-pdf==0.2.0.dev1
django-grappelli==2.10.1
django-import-export==1.1.0
django-jet==1.0.8
django-modeladmin-reorder==0.2
django-rest-auth==0.9.1
djangorestframework==3.6.3
drfapikey==0.0.3
drfdocs==0.0.11
enum34==1.1.6
et-xmlfile==1.0.1
fcm-django==0.2.11
futures==3.2.0
html5lib==1.0.1
idna==2.5
isort==4.3.4
itypes==1.1.0
jdcal==1.4
Jinja2==2.9.6
lazy-object-proxy==1.3.1
lxml==4.3.0
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.2
odfpy==1.3.6
olefile==0.44
openapi-codec==1.3.1
openpyxl==2.5.9
pdfrw==0.4
Pillow==4.2.1
psycopg2==2.7.1
psycopg2-binary==2.7.6.1
pycparser==2.19
pyfcm==1.3.1
pylint==1.9.3
PyPDF2==1.26.0
Pyphen==0.9.5
python-openid==2.2.5
pytz==2017.2
PyYAML==3.13
raven==6.1.0
reportlab==3.5.10
requests==2.18.1
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.10.0
tablib==0.12.1
tinycss==0.4
tinycss2==0.6.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.21.1
WeasyPrint==0.42.3
webencodings==0.5.1
wrapt==1.10.11
xhtml2pdf==0.2.3
xlrd==1.1.0
xlwt==1.3.0
